+++ LEARN_JS_TASKS/HomeworksJS/For-While.js
