+++ LEARN_JS_TASKS/HomeworksJS/Functions.js
