+++ LEARN_JS_TASKS/HomeworksJS/If.js
