+++ LEARN_JS_TASKS/HomeworksJS/Inheritance.js
