+++ LEARN_JS_TASKS/HomeworksJS/Logical operators.js
