+++ LEARN_JS_TASKS/HomeworksJS/Logical_bit_operators.js
