+++ LEARN_JS_TASKS/HomeworksJS/Objects.js
