+++ LEARN_JS_TASKS/HomeworksJS/Switch.js
