+++ LEARN_JS_TASKS/HomeworksJS/Try_Catch.js
