+++ LEARN_JS_TASKS/HomeworksJS/Types.js
