+++ LEARN_JS_TASKS/HomeworksJS/Variables.js
